refractor: retire high-priority structural debt in router backend hotspots (#1969)
