+++ tools/agent/structure-rules.yaml
@@ -75,8 +75,6 @@ structure_exceptions:
     file_checks: relaxed
   - paths:
       - src/semantic-router/pkg/config/config_test.go
-      - src/semantic-router/pkg/dsl/compiler.go
-      - src/semantic-router/pkg/dsl/decompiler.go
       - src/semantic-router/pkg/dsl/dsl_test.go
     file_checks: relaxed
     function_checks: relaxed
@@ -176,10 +174,7 @@ structure_exceptions:
     file_checks: relaxed
     function_checks: relaxed
   - paths:
-      - src/semantic-router/pkg/memory/milvus_store.go
       - src/semantic-router/pkg/vectorstore/milvus_backend.go
-    file_checks: relaxed
-    function_checks: relaxed
   - paths:
       - src/semantic-router/pkg/classification/hallucination_detector.go
       - src/semantic-router/pkg/classification/mcp_classifier.go
